May7

+++ src/main/java/com/hmkurth/ApiLocation/Location.java
@@ -2,6 +2,9 @@ package com.hmkurth.ApiLocation;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+/**
+ * The type Location.
+ */
 public class Location{
 
 	@JsonProperty("lng")
@@ -10,30 +13,60 @@ public class Location{
 	@JsonProperty("lat")
 	private double lat;
 
-	public String getName() {
+    /**
+     * Gets name.
+     *
+     * @return the name
+     */
+    public String getName() {
 		return name;
 	}
 
-	public void setName(String name) {
+    /**
+     * Sets name.
+     *
+     * @param name the name
+     */
+    public void setName(String name) {
 		this.name = name;
 	}
 
 	@JsonProperty("name")
 	private String name;
 
-	public void setLng(double lng){
+    /**
+     * Set lng.
+     *
+     * @param lng the lng
+     */
+    public void setLng(double lng){
 		this.lng = lng;
 	}
 
-	public double getLng(){
-		return lng;
+    /**
+     * Get lng float.
+     *
+     * @return the float
+     */
+    public float getLng(){
+		return (float) lng;
 	}
 
-	public void setLat(double lat){
+    /**
+     * Set lat.
+     *
+     * @param lat the lat
+     */
+    public void setLat(double lat){
 		this.lat = lat;
 	}
 
-	public double getLat(){
+    /**
+     * Get lat double.
+     *
+     * @return the double
+     */
+    public double getLat(){
 		return lat;
 	}
 }
